requirements/base.txt
Django>=5.0.4,<5.1
django-environ>=0.11.2,<0.12
djangorestframework>=3.15.1,<3.16
django-cors-headers>=4.3.1,<4.4
django-filter>=24.2,<24.3
django-autoslug>=1.9.9,<1.10
django-countries>=7.6.1,<7.7
django-phonenumber-field>=7.3.0,<7.4
phonenumbers>=8.13.36,<8.14
drf-yasg>=1.21.7,<1.22
Pillow>=10.3.0,<10.4
argon2-cffi>=23.1.0,<23.2
pytz>=2024.1,<2024.2
redis>=5.0.4,<5.1
celery>=5.4.0,<5.5
flower>=2.0.1,<2.1
django-celery-email>=3.0.0,<3.1
django-allauth>=0.63.1,<0.64
djangorestframework-simplejwt>=5.3.1,<5.4
dj-rest-auth>=6.0.0,<6.1
requests>=2.32.1,<2.33
